+++ src/components/SubSel.js
@@ -1,4 +1,8 @@
 import React, { useState, useEffect, useRef } from "react";
+import { useLoggedIn } from "../stores/loggedIn.js";
+
+// stores
+import { useSnoo } from "./../stores/snoo.js";
 
 // css
 import "./SubSel.css";
@@ -8,9 +12,12 @@ export const SubredditSelect = ({
 	subreddit,
 	closeSelectMenu
 }) => {
+	// todo: add react-hook-forms !
 	// todo: add auto complete and make it a full power menu.
 	const subSel = useRef();
 	const [inputSub, setInputSub] = useState(`r/${subreddit}`);
+
+	const loggedIn = useLoggedIn((s) => s.loggedIn);
 
 	// idk why but this doesnt work when the input field is focused.
 	// https://github.com/greena13/react-hotkeys/issues/100
@@ -36,6 +43,8 @@ export const SubredditSelect = ({
 				ref={subSel}
 				value={inputSub}
 				onKeyDown={(e) => {
+					// this is imp because on pressing `j`,`k` scroll handlers might get triggered.
+					e.stopPropagation();
 					e.key === "Enter" && sel_subreddit(inputSub);
 					if (e.key === "Escape") {
 						subSel.current.blur();
@@ -49,6 +58,38 @@ export const SubredditSelect = ({
 				}}
 				placeholder={`r/${subreddit}`}
 			/>
+			{loggedIn && (
+				<SuggestionsList
+					query={inputSub}
+					sel_subreddit={sel_subreddit}
+				/>
+			)}
 		</div>
+	);
+};
+
+const SuggestionsList = ({ query, sel_subreddit }) => {
+	const snoo = useSnoo((s) => s.snoo);
+	// stores suggested sub names, should i store results to avoid fetching again & again ?
+	const [suggestions, setSuggestions] = useState([]);
+	const time = 500;
+	// ok so this has to be a debounce thing
+	useEffect(() => {
+		if (query === "r/") return null;
+		const timer = setTimeout(() => {
+			snoo.searchSubredditNames({ query }).then(setSuggestions);
+		}, time);
+		return () => {
+			clearTimeout(timer);
+		};
+		//http://localhost:3000/auth_redirect#access_token=-iAbSbIsOBvoTE5llVYXnx09Khv5XZg
+	}, [query]);
+
+	return (
+		<ul>
+			{suggestions.map((i) => (
+				<li onClick={() => sel_subreddit("r/" + i)}>{i}</li>
+			))}
+		</ul>
 	);
 };