+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 ## CHANGELOG
 
-1. added scrolling with 'j' and 'k', had to remove overall scroll-behavior: smooth !
-2. added profile pics to posts !
-3. fixed the snoo bug ! now logged in users can view different listings !
+1. some refactoring.
+2. added subreddit names debounced search !